+++ lib/govuk_publishing_components.rb
@@ -43,8 +43,6 @@ require "govuk_publishing_components/app_helpers/brand_helper"
 require "govuk_publishing_components/app_helpers/environment"
 require "govuk_publishing_components/app_helpers/asset_helper"
 
-require "govuk_publishing_components/railtie" if defined?(Rails::Railtie)
-
 # Add i18n paths and views for usage outside of a Rails app
 I18n.load_path.unshift(
   *Dir.glob(File.expand_path("config/locales/*.yml", GovukPublishingComponents::Config.gem_directory)),
@@ -54,6 +52,10 @@ ActiveSupport.on_load(:action_controller) do
   ActionController::Base.append_view_path(
     File.expand_path("app/views", GovukPublishingComponents::Config.gem_directory),
   )
+end
+
+ActiveSupport.on_load(:action_view) do
+  include GovukPublishingComponents::AppHelpers::AssetHelper
 end
 
 module GovukPublishingComponents